Extract sampling stuff to its own file:
