+++ src/index.ts
@@ -1,8 +1,7 @@
 import * as appInsights from 'applicationinsights'
-import { EventTelemetry, DependencyTelemetry, ExceptionTelemetry, MetricTelemetry, RequestTelemetry, TraceTelemetry, Envelope } from 'applicationinsights/out/Declarations/Contracts'
-import { samplingTelemetryProcessor } from 'applicationinsights/out/TelemetryProcessors'
-import Context = require('applicationinsights/out/Library/Context')
+import { EventTelemetry, DependencyTelemetry, ExceptionTelemetry, MetricTelemetry, RequestTelemetry, TraceTelemetry } from 'applicationinsights/out/Declarations/Contracts'
 const clientKey = (process.env.APPINSIGHTS_INSTRUMENTATIONKEY || "fake")
+import { addSamplingRulesByUrl, RulesDictonary } from './samplingRulesByUrl'
 
 appInsights.setup(clientKey)
     .setAutoDependencyCorrelation(<boolean>(process.env.AI_AUTO_DEPENDENCY_CORRELATE === 'false' ? false : true))
@@ -55,24 +54,6 @@ export function measureDependency (marker: IMarker, data = '', success = true): 
   trackDependency(telemetry)
 }
 
-interface RulesDictonary {
-  // example: '/api/v3/calculation': 50
-  [key: string]: number
-}
-
-const samplingRulesByUrl = (envelope: Envelope, context: any, rulesDictionary: RulesDictonary = {}) => {
-  // if it's not an http request, use the regular sampling processor
-  if (!context['http.RequestOptions']) {
-    return samplingTelemetryProcessor(envelope, { correlationContext: context })
-  }
-  // otherwise, use the rules dictionary to determine how much to sample
-  const pathname = (context['http.RequestOptions'].uri || {}).pathname
-  const samplingRate = rulesDictionary[pathname] || aiClient.config.samplingPercentage
-
-  // if false returned from a telemetry processor, the data will not be sent
-  return samplingRate >= (Math.random() * 100)
-} 
-
-export const addSamplingRulesByUrl = (rulesDictionary: RulesDictonary) => {
-  aiClient.addTelemetryProcessor((envelope: Envelope, context: Context) => samplingRulesByUrl(envelope, context, rulesDictionary))
+export function samplingRulesByUrl (rulesDictionary: RulesDictonary) {
+  addSamplingRulesByUrl(rulesDictionary, aiClient)
 }
